+++ test/test_clyap.jl
@@ -139,12 +139,16 @@ Ty == Float64 ? reltol = eps(float(100)) : reltol = eps(100*n*one(Ty))
 @time x = lyapc(ar,qc);
 @test norm(ar*x+x*ar'+qc)/norm(x)/norm(ar) < reltol
 
-#if Ty != Float32
-  # Fix for missing strsyl3 in OpenBLAS   
 @time x = lyapc(ar,Qnh);
 @test norm(ar*x+x*ar'+Qnh)/norm(x)/norm(ar)  < reltol
 
+@time x = lyapc(ar,Qnh,blocksize=5);
+@test norm(ar*x+x*ar'+Qnh)/norm(x)/norm(ar)  < reltol
+
 @time x = lyapc(ar',Qnh);
+@test norm(ar'*x+x*ar+Qnh)/norm(x)/norm(ar)  < reltol
+
+@time x = lyapc(ar',Qnh,blocksize=5);
 @test norm(ar'*x+x*ar+Qnh)/norm(x)/norm(ar)  < reltol
 
 @time x = lyapc(ars,Qnh);
@@ -155,7 +159,13 @@ Ty == Float64 ? reltol = eps(float(100)) : reltol = eps(100*n*one(Ty))
 @time x = lyapc(ac,qnh);
 @test norm(ac*x+x*ac'+qnh)/norm(x)/norm(ac)  < reltol
 
+@time x = lyapc(ac,qnh,blocksize=5);
+@test norm(ac*x+x*ac'+qnh)/norm(x)/norm(ac)  < reltol
+
 @time x = lyapc(ac',qnh);
+@test norm(ac'*x+x*ac+qnh)/norm(x)/norm(ac)  < reltol
+
+@time x = lyapc(ac',qnh,blocksize=5);
 @test norm(ac'*x+x*ac+qnh)/norm(x)/norm(ac)  < reltol
 
 
@@ -213,6 +223,8 @@ c = rand(Ty,n,n)+im*rand(Ty,n,n)
 #qc = c'*c
 qc = Matrix(Hermitian(c'*c));
 Qr = real(qc)
+qnh = c;
+Qnh = real(qnh);
 Ty == Float64 ? reltol = eps(float(100)) : reltol = eps(100*n*one(Ty))
 
 
@@ -221,6 +233,12 @@ Ty == Float64 ? reltol = eps(float(100)) : reltol = eps(100*n*one(Ty))
 
 @time x = lyapc(ac,ec,qc,blocksize=5);
 @test norm(ac*x*ec'+ec*x*ac'+qc)/norm(x)/norm(ac)/norm(ec) < reltol
+
+@time x = lyapc(ac,ec,qnh);
+@test norm(ac*x*ec'+ec*x*ac'+qnh)/norm(x)/norm(ac)/norm(ec) < reltol
+
+@time x = lyapc(ac,ec,qnh,blocksize=5);
+@test norm(ac*x*ec'+ec*x*ac'+qnh)/norm(x)/norm(ac)/norm(ec) < reltol
 
 β = 3
 @time x = lyapc(ac,β*I,qc);
@@ -248,6 +266,13 @@ Ty == Float64 ? reltol = eps(float(100)) : reltol = eps(100*n*one(Ty))
 @time x = lyapc(ac',ec',qc,blocksize=5);
 @test norm(ac'*x*ec+ec'*x*ac+qc)/norm(x)/norm(ac)/norm(ec)  < reltol
 
+@time x = lyapc(ac',ec',qnh);
+@test norm(ac'*x*ec+ec'*x*ac+qnh)/norm(x)/norm(ac)/norm(ec)  < reltol
+
+@time x = lyapc(ac',ec',qnh,blocksize=5);
+@test norm(ac'*x*ec+ec'*x*ac+qnh)/norm(x)/norm(ac)/norm(ec)  < reltol
+
+
 β = (1+im);
 @time x = lyapc(ac',β*I,qc);
 @test norm(ac'*x*β+β'*x*ac+qc)/norm(x)/norm(ac)/norm(ec)  < reltol
@@ -268,6 +293,12 @@ Ty == Float64 ? reltol = eps(float(100)) : reltol = eps(100*n*one(Ty))
 @time x = lyapc(ar,er,Qr,blocksize=5);
 @test norm(ar*x*er'+er*x*ar'+Qr)/norm(x)/norm(ar)/norm(er)  < reltol
 
+@time x = lyapc(ar,er,Qnh);
+@test norm(ar*x*er'+er*x*ar'+Qnh)/norm(x)/norm(ar)/norm(er)  < reltol
+
+@time x = lyapc(ar,er,Qnh,blocksize=5);
+@test norm(ar*x*er'+er*x*ar'+Qnh)/norm(x)/norm(ar)/norm(er)  < reltol
+
 @time x = lyapc(ar,er,qc);
 @test norm(ar*x*er'+er*x*ar'+qc)/norm(x)/norm(ar)/norm(er)  < reltol
 
@@ -277,6 +308,11 @@ Ty == Float64 ? reltol = eps(float(100)) : reltol = eps(100*n*one(Ty))
 @time x = lyapc(ar',er',Qr,blocksize=5);
 @test norm(ar'*x*er+er'*x*ar+Qr)/norm(x)/norm(ar)/norm(er) < reltol
 
+@time x = lyapc(ar',er',Qnh);
+@test norm(ar'*x*er+er'*x*ar+Qnh)/norm(x)/norm(ar)/norm(er) < reltol
+
+@time x = lyapc(ar',er',Qnh,blocksize=5);
+@test norm(ar'*x*er+er'*x*ar+Qnh)/norm(x)/norm(ar)/norm(er) < reltol
 
 @time x = lyapc(ar',er,Qr);
 @test norm(ar'*x*er'+er*x*ar+Qr)/norm(x)/norm(ar)/norm(er) < reltol